+++ src/main/java/mchorse/bbs_mod/utils/keyframes/KeyframeSegment.java
@@ -15,7 +15,7 @@ public class KeyframeSegment <T>
 
     public Keyframe<T> preA;
     public Keyframe<T> postB;
-    public int duration;
+    public float duration;
     public float offset;
     public float x;
 
@@ -60,9 +60,9 @@ public class KeyframeSegment <T>
     {
         int forcedDuration = this.a.getDuration();
 
-        this.duration = forcedDuration > 0 ? forcedDuration : (int) (this.b.getTick() - this.a.getTick());
+        this.duration = forcedDuration > 0 ? forcedDuration : this.b.getTick() - this.a.getTick();
         this.offset = ticks - this.a.getTick();
-        this.x = this.duration == 0 ? 0F : this.offset / (float) this.duration;
+        this.x = this.duration == 0 ? 0F : this.offset / this.duration;
     }
 
     public T createInterpolated()
